refactor services and controllers

+++ src/main/java/cl/gbarrera/demo/model/Product.java
@@ -11,13 +11,13 @@ import lombok.*;
 @AllArgsConstructor
 public class Product {
 
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
 
-  @Column(nullable = false)
-  private String name;
+    @Column(nullable = false)
+    private String name;
 
-  @Column(nullable = false)
-  private Integer price;
+    @Column(nullable = false)
+    private Long price;
 }
